Added German translation.
Provided a German translation for the fixers manifesto. The main problem
with the translation is that the word "fixer" cannot be easily
translated into German directly. A literal translation is not possible
because the German word "Fixer" means a drug addict. As there is no
German noun for a person reparing things, I forged a term from the
German verb for repairing "reparieren" und the suffix for persons doing
sometihing "er".